pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gitmentor"
version = "1.0.0"
description = "Autonomous Git Agent & Code Steward powered by LangGraph"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "GitMentor Contributors" }
]
license = { text = "MIT" }
dependencies = [
  "langchain-core>=0.2.3",
  "langchain-google-genai>=1.0.5",
  "langgraph>=0.2.20",
  "langchain>=0.2.12",
  "python-dotenv>=1.0.1",
  "GitPython>=3.1.43",
  "PyYAML>=6.0.2",
  "rich>=13.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "black>=24.8.0"
]

[project.scripts]
gm = "main:main"
gitmentor = "main:main"

[tool.setuptools]
py-modules = ["main"]
packages = { find = { where = ["."] } }